Export and use Tables.changeCellHandler's functionality
This patch moves the function body of the changeCellHandler out of the Tables
class and makes it available for import. Calling ActionCreator.changeCell will
use it as before.

The import returns an thenable, so it can be used in situations where work
needs to be done after the data base change has occurred (or failed.