Replace CellMeasurer with custom solution
In a NxM table, CellMeasurer will require each cell in a row to
re-render up to N-1 times to make sure it has the correct height, no
matter if the cell is actually visible.
While this worked, it would completely negate the effects of using
react-virtualized, slowing the browser down to a crawl on tables
with many columns.
This patch will just look if id a row's id is within the array of
expanded rows and then immediately set its height to one of the
possible heights, effectively replacing the O(N^2) pre-rendering with
two array lookups, then signalling a necessary grid resize to
react-virtualize MultiGrid.

The same technique can be used to resize columns simultaneously,
something that CellMeasurer does not support at all.