Fix concats failing on empty cells in table 147/column/3
For some unknown reason, deleting all links from a row in table 147,
column 3 of the test centre causes the map to crash, even although the
called function `flattenAndTranslate` uses a default-value of `[]`.

This patch fixed the crash, but further investigation is needed. The
column type validates specifications, as it is a link column with
multi language support, but there is also a similarly bad link column
in table 75; setting it as `id: true` and clearing it does not provoke
that error.