Refactor date edit cell to display changes immediately
Now choosing values in the date picker will display selected values
immediately, while changes get only saved when the cell unmounts.

This requires a somewhat ugly hack: We mirror the current value state,
which is used for display, in a mutable ref. Now the `useEffect` hook
which saves the value can be modified to run only on unmount, ignoring
changes of the display value (indicated by the empty array as second
parameter), and saving the content of the mutable ref.