Fix error when setting annotations in row with annotation
When trying to set an annotation in a row where any cell had already
annotations, `existingAnnotation` did not return a proper falsy value
but an empty array, so downstream toggle functions assumed they should
delete the annotation and sent requests to `undefined`.

Fixed a downstream error in `rows.js`, which was supposed to be worked
around by returning empty array instead of `nil`.