Hide history overlay option for group and concat columns
Group and concat columns are only virtual columns. The backend
generates them on the fly from other columns' values, thus it is not
feasible to generate meaningful diff values for them; their only event
will be "row_created".
To account for this, we don't allow the user to open the revision
history overlay for those kind of cells.