Fix adding new foreign rows crashing GRUD
The initial render of the EntityView overlay that opens after a new
foreign row as created from the LinkOverlay can in some cases include
cells with empty fields. We provide empty object fallback values for
this case, thus allowing the overlay to recover on subsequent
renders (e.g. when row was successfully retrieved from the server).