Fix: Update display values for foreign entity views
The display value action triggered by the foreignEntityView passed the
new display values in a wrong format, so the cache was never updated.

The tableView reducer always preferred existing cache values over
new ones to avoid duplicating them. This was to ensure that fetching
displayValues for link concats would not overwrite all other columns
in a foreign row. Now we explicitly check, if a displayValue update
delivers only a preview or a full row, and always keep the full
row. We can do this safely, because after going into the state,
modified displayValues will be synched with the server and need no
extra fetching.