Initially show all translations in popup
When opening the translation popup for the first time, no translations
were selected. This would result in the translation popup being
completely empty, which would confuse users, as it was hard to see how
to show any other languages.
To fix this, we invert the behaviour to showing all available
translations as default, leaving the user with the choice to hide
unwanted ones.