Handle badly defined link columns
The API does not forbid to set link columns to have `multilanguage:
true`, although link values are always single language arrays. This
would trip the `needsUpdate` check, as no language keys get changed in
these cases. This commit will make the multilanguage check more
strict, by checking the `multilalnguage` flag and then performing a
plausibility check on the provided value.

A value suffices as multi language if it is either a plain object or
`nil`.