Add fresh rows to cache
For performance reasons, we avoid re-fetching available link
candidates, but create an array of links from  the cell's value and
the available rows when opening the overlay (cells with cardinality
will re-fetch available candidates then). On updates, cells are taken
from the array and sorted to "linked" or "unlinked".
Newly created rows will appear in the cell's value only; if we unlink
them without closing the overlay, they will not be available from
either the initial foreinRows request or the cell's value, thus
disappear from the overlay for good.

To solve this without needing to re-fetch foreign rows, we add new
rows to the foreignRows array once they were successfully created.