Fix attachment history always showing "item was deleted"
We also add current files' display values to the shared
link/attachment diff visualiser. We don't check if files have been
deleted since, because to do so we would need to fetch all files and
check for 404 responses, also we can not retrieve name changes this
way. This should not matter too much, as a file is either there, or it
isn't.