Add visual feedback to text subject to linking
When we click "add link", we lose the visible selection in the text
editor. To let the user know, where the link will be set we exploit
the "underline" style attribute, which exists in richtext but not in
markdown and restyle it to highlight the text which will receive the
link attribute.